feat: 네트워크 요청 보안 설정 및 네트워크 연결 여부에 따른 오류 수정

+++ app/src/main/java/com/cupofcoffee/ui/settings/SettingsViewModel.kt
@@ -13,6 +13,7 @@ import com.cupofcoffee.data.DataResult
 import com.cupofcoffee.data.DataResult.Companion.error
 import com.cupofcoffee.data.DataResult.Companion.loading
 import com.cupofcoffee.data.DataResult.Companion.success
+import com.cupofcoffee.data.repository.CommentRepositoryImpl
 import com.cupofcoffee.data.repository.MeetingRepositoryImpl
 import com.cupofcoffee.data.repository.PlaceRepositoryImpl
 import com.cupofcoffee.data.repository.PreferencesRepositoryImpl
@@ -26,6 +27,7 @@ class SettingsViewModel(
     private val userRepositoryImpl: UserRepositoryImpl,
     private val meetingRepositoryImpl: MeetingRepositoryImpl,
     private val placeRepositoryImpl: PlaceRepositoryImpl,
+    private val commentRepositoryImpl: CommentRepositoryImpl,
     private val preferencesRepositoryImpl: PreferencesRepositoryImpl,
     private val networkUtil: NetworkUtil
 ) : ViewModel() {
@@ -49,7 +51,7 @@ class SettingsViewModel(
         }
     }
 
-    fun convertIsAutoLogin(){
+    fun convertIsAutoLogin() {
         viewModelScope.launch {
             preferencesRepositoryImpl.toggleAutoLogin()
         }
@@ -68,6 +70,7 @@ class SettingsViewModel(
             deleteMeeting(meetingEntry)
             deleteMadeMeetingsInPlace(meetingEntry.meetingModel.placeId, meetingId)
         }
+        deleteComments(uid)
         deleteUser(user)
     }
 
@@ -85,10 +88,15 @@ class SettingsViewModel(
     private suspend fun deleteMadeMeetingsInPlace(placeId: String, meetingId: String) {
         val placeEntry = placeRepositoryImpl.getPlaceById(placeId, isConnected())!!
         placeEntry.placeModel.meetingIds.remove(meetingId)
-        if (placeEntry.placeModel.meetingIds.isEmpty()) {
-            placeRepositoryImpl.delete(placeEntry)
-        } else {
-            placeRepositoryImpl.update(placeEntry)
+        if (placeEntry.placeModel.meetingIds.isEmpty()) placeRepositoryImpl.delete(placeEntry)
+        else placeRepositoryImpl.update(placeEntry)
+    }
+
+    private suspend fun deleteComments(userId: String) {
+        val commentIdsByUserId =
+            commentRepositoryImpl.getCommentsByUserId().filterValues { it.userId == userId }.keys
+        commentIdsByUserId.forEach { id ->
+            commentRepositoryImpl.delete(id)
         }
     }
 
@@ -103,6 +111,7 @@ class SettingsViewModel(
                     userRepositoryImpl = CupOfCoffeeApplication.userRepository,
                     meetingRepositoryImpl = CupOfCoffeeApplication.meetingRepository,
                     placeRepositoryImpl = CupOfCoffeeApplication.placeRepository,
+                    commentRepositoryImpl = CupOfCoffeeApplication.commentRepository,
                     preferencesRepositoryImpl = CupOfCoffeeApplication.preferencesRepositoryImpl,
                     networkUtil = CupOfCoffeeApplication.networkUtil
                 )
